create CombDistribution
